Move Preprocessor implem to non anonymous class
To make it easier to tweak it when using Ammonite as a library
